unify fetching of a single property
