+++ src/routes/category-property/[id]/+page.svelte
@@ -4,13 +4,4 @@
 	let { data } = $props()
 </script>
 
-<PropertyPage
-	type="category"
-	property={data.property}
-	related_properties={data.related_properties}
-	relevant_implications={data.relevant_implications}
-	examples={data.examples}
-	counterexamples={data.counterexamples}
-	unknown_structures={data.unknown_categories}
-	undecidable_structures={data.undecidable_categories}
-/>
+<PropertyPage type="category" {...data} />
